# api/docs/swagger.yaml
definitions:
...
  template_variables.EmptyResponse:
    type: object
  template_variables.FailureResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  var_user_service.LoginExistsRequest:
    properties:
      login:
        type: string
    type: object
  var_user_service.LoginExistsResponse:
    properties:
      exist:
        type: boolean
    type: object
  var_user_service.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  var_user_service.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      verified:
        type: boolean
    type: object
...
  var_user_service.UpdatePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /v1/login:
    post:
      consumes:
      - application/json
      description: API to singin
      parameters:
      - description: login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/var_user_service.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/var_user_service.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: login
      tags:
      - auth
  /v1/login-exists:
    post:
      consumes:
      - application/json
      description: API to singin
      parameters:
      - description: Login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/var_user_service.LoginExistsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/var_user_service.LoginExistsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: login
      tags:
      - auth
  /v1/login-refresh:
    post:
      consumes:
      - application/json
      description: API to get your access token with refresh
      parameters:
      - description: refresh_token
        in: query
        name: refresh_token
        required: true
        type: string
      - description: is_applicant
        in: query
        name: is_applicant
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/var_user_service.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: if access-token expired, get your access token with refresh
      tags:
      - auth
...
  /v1/update-password:
    post:
      consumes:
      - application/json
      description: API to update user password
      parameters:
      - description: Update password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/var_user_service.UpdatePassword'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/template_variables.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      security:
      - ApiKeyAuth: []
      summary: update user password
      tags:
      - auth
  /v1/update-password/{user_id}:
    post:
      consumes:
      - application/json
      description: API to update user password
      parameters:
      - description: User id
        in: path
        name: user_id
        required: true
        type: string
      - description: Update password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/var_user_service.UpdatePassword'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/template_variables.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      security:
      - ApiKeyAuth: []
      summary: update user password
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"